// src/model/programacao.ts
import { Palestrante } from './palestrante';

export class Programacao {
    idReferencia: string;
    titulo: string;
    descricao: string;
    imagem: string;
    conteudo: string;
    tipo: string;
    data: string;
    horario: string;
    // palestrante: Palestrante;
    palestrante: string;
}